# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/ContentLibrary/GetAllLibraries": {
      "post": {
        "tags": [
          "ContentLibrary"
        ],
        "operationId": "ContentLibrary_GetAllLibraries",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryFilter"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryFilter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryFilter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryFilter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult_LibraryDto"
                }
              }
            }
          }
        }
      }
    },
    "/ContentLibrary/GetUserLibraries": {
      "post": {
        "tags": [
          "ContentLibrary"
        ],
        "operationId": "ContentLibrary_GetUserLibraries",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoLibraryFilterValueTuple"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoLibraryFilterValueTuple"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoLibraryFilterValueTuple"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDtoLibraryFilterValueTuple"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResult_LibraryDto"
                }
              }
            }
          }
        }
      }
    },
    "/ContentLibrary/GetLibraryById": {
      "get": {
        "tags": [
          "ContentLibrary"
        ],
        "operationId": "ContentLibrary_GetLibraryById",
        "parameters": [
          {
            "name": "libraryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              }
            }
          }
        }
      }
    },
    "/ContentLibrary/CreateNewLibrary": {
      "post": {
        "tags": [
          "ContentLibrary"
        ],
        "operationId": "ContentLibrary_CreateNewLibrary",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ContentLibrary/UpdateLibraryById": {
      "put": {
        "tags": [
          "ContentLibrary"
        ],
        "operationId": "ContentLibrary_UpdateLibraryById",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/ContentLibrary/DeleteLibraryById": {
      "delete": {
        "tags": [
          "ContentLibrary"
        ],
        "operationId": "ContentLibrary_DeleteLibraryById",
        "parameters": [
          {
            "name": "libraryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/File/UploadFile": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "File_UploadFile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/File/GetFileById": {
      "get": {
        "tags": [
          "File"
        ],
        "operationId": "File_GetFileById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/File/DeleteFile": {
      "delete": {
        "tags": [
          "File"
        ],
        "operationId": "File_DeleteFile",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/File/UploadChunk": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "File_UploadChunk",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadChunkDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadChunkDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadChunkDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UploadChunkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/File/FinalizeUpload": {
      "post": {
        "tags": [
          "File"
        ],
        "operationId": "File_FinalizeUpload",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/FinaliseUploadDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinaliseUploadDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FinaliseUploadDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FinaliseUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinaliseUploadDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LibraryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LibraryFilter": {
        "type": "object",
        "properties": {
          "libraryIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "libraryName": {
            "type": "string",
            "nullable": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PagedResult_LibraryDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UploadChunkDto": {
        "type": "object",
        "properties": {
          "chunk": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "uploadId": {
            "type": "string",
            "nullable": true
          },
          "chunkNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UserDtoLibraryFilterValueTuple": {
        "type": "object",
        "additionalProperties": false
      },
...
    }
  }
}